+++ CHANGELOG.md
@@ -1,7 +1,20 @@
 # Change Log
 
-## [4.4.0](https://github.com/plivo/plivo-python/tree/v4.4.0) (2020-04-29)
+
+## [4.7.0](https://github.com/plivo/plivo-python/tree/v4.7.0) (2020-04-29)
 - Add V3 signature helper functions.
+
+## [4.6.1](https://github.com/plivo/plivo-python/tree/v4.6.1) (2020-03-31)
+- Add Nonetype check for API requests.
+
+## [4.6.0](https://github.com/plivo/plivo-python/tree/v4.6.0) (2020-03-31)
+- Add application cascade delete support.
+
+## [4.5.0](https://github.com/plivo/plivo-python/tree/v4.5.0) (2020-03-30)
+- Add Tollfree support for Powerpack
+
+## [4.4.0](https://github.com/plivo/plivo-python/tree/v4.4.0) (2020-03-27)
+- Add post call quality feedback API support.
 
 ## [4.3.8](https://github.com/plivo/plivo-python/tree/v4.3.8) (2020-02-25)
 - Add Media support